third_party/WebKit/LayoutTests/http/tests/fetch/chromium/close-worker-while-requesting.html

<!doctype html>
<script src = "/resources/testharness.js"></script>
<script src = "/resources/testharnessreport.js"></script>
<script>
promise_test(() => {
    const SCRIPT = '/fetch/chromium/resources/close-worker-while-requesting.js';
    worker = new Worker(SCRIPT);
    return new Promise(resolve => {
        worker.onmessage = message => {
            assert_equals(message.data, 'PASS');
            resolve();
          };
      });
  });
</script>
<p>This test checks that Blink doesn't crash when closing a worker that has
active fetches.</p>
